requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
pydantic==2.9.2
pymongo==4.10.1
python-dotenv==1.0.1
